Refactor code to use an object to store choices
By storing the choices as object properties, it allows the the property
values to be set to the loosing object, or the beaten object according
to the rock paper scissors rules. This will make it very easy to
determine winner when that functionality gets implemented later. The
functions for getting choices have been updated to reflect this change,
using the Object.keys(object) method to get the attribute names in
array format.