Add the ability to play rounds with the computer
Create the function playRound(humanChoice, computerChoice) to play
a round with the computer. The winner is calculated using the choices
object, which contains the loosing value for each possible choice. Also
adds two variables to keep track of how many wins the computer and the
user have. These are adjusted accordingly in the playRound function.
Added some test code to play rounds infinately.