Refactor code and implement playGame function
Refactor the playRound function and score variables to within a new
function, playGame(rounds = 5). This function will play a number of
rounds, defaulting to 5. It handles invalid input for user, and only
advances a round if the round was not a draw. To determine this, the
playRound() function returns true for valid rounds, and false for draws.

There was also a small typo in the playRound function where
computerChoice variable was getting incremented instead of
computerScore.